requirements.txt
alembic==1.15.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
SQLAlchemy==2.0.37
Werkzeug==3.1.3
